new GetOneClient method

+++ BankAccount.Backend.Domain/ViewModel/UpdateAccountViewModel.cs
@@ -6,7 +6,7 @@ namespace BankAccount.Backend.Domain.ViewModel
     {
         public int Id { get; set; }
         public string? AccountNumber { get; set; }
-        public float? Balance { get; set; }
+        public decimal? Balance { get; set; }
         public AccountType? AccountType { get; set; }
         public int? CreditLimit { get; set; }
         public int? ClientID { get; set; }
